+++ src/DP/TSPModel.cuh
@@ -10,6 +10,6 @@ namespace DP
             __host__ void makeRoot(OP::TSPProblem const * problem,  TSPState* root);
             __device__ void makeNextState(OP::TSPProblem const * problem, TSPState const * state, int value, int cost, TSPState* nextState);
             __device__ void mergeNextState(OP::TSPProblem const * problem,  TSPState const * state, int value, TSPState* nextState);
-            __device__ void calcCosts(OP::TSPProblem const * problem, unsigned int level, TSPState const * state, int * costs);
+            __device__ void calcCosts(OP::TSPProblem const * problem, unsigned int level, TSPState const * state, int16_t * costs);
     };
 }
